・イメージ画像の表記修正 ・テストアカウントは削除できないように修正

+++ app/Http/Controllers/MyPage/Register_info/DeleteProcessController.php
@@ -10,7 +10,11 @@ use Illuminate\Support\Facades\Auth;
 class DeleteProcessController extends Controller
 {
     public function __invoke(Request $request){
-        User::find(Auth::user()->id)->delete();
-        return redirect('/msg')->with('title','退会処理完了')->with('msg','退会処理が完了しました。');
+        if(Auth::user()->email=='test@example.com'){
+            return redirect('/msg')->with('title','エラー')->with('msg','テストアカウントは削除できません。');
+        }else{
+            User::find(Auth::user()->id)->delete();
+            return redirect('/msg')->with('title','退会処理完了')->with('msg','退会処理が完了しました。');
+        }
     }
 }
